Week 4 opdr 2 af

+++ src/main/java/Betaalwijze.java
@@ -1,13 +1,15 @@
+import java.math.*;
+
 public abstract class Betaalwijze {
 
-    protected double saldo;
+    protected BigDecimal saldo;
 
     /**
      * Methode om krediet te initialiseren
      *
      * @param saldo
      */
-    public void setSaldo(double saldo) {
+    public void setSaldo(BigDecimal saldo) {
         this.saldo = saldo;
     }
 
@@ -17,5 +19,5 @@ public abstract class Betaalwijze {
      * @param tebetalen
      * @return Boolean om te kijken of er voldoende saldo is
      */
-    public abstract boolean betaal(double tebetalen);
+    public abstract boolean betaal(BigDecimal tebetalen);
 }
